"Update cart functionality, fix stock calculation, and adjust cookie handling in Passport and utils"
This commit includes the following changes:

1. In `passport.config.js`, the console log for the `req` object has been commented out.
2. In `carts.controller.js`, the `addProductToCart` function has been updated to parse the `quantity` from the request body, get the `productId` and `cartId` from the user object, and log an empty line for better readability.
3. In `purchaseCart` function, the Promise.all loop has been updated to use `cart.products` instead of `cart[0].products`.
4. In `CartsDB.js`, the `deleteProduct` function has been updated to use `cart.products` instead of `cart[0].products`.
5. In `ProductDTO.js`, the `stock` property type has been changed from `parseFloat` to `parseInt`.
6. In `utils.js`, the console log for `cookiesString` has been commented out.

+++ src/dto/ProductDTO.js
@@ -6,7 +6,7 @@ export default class ProductDTO {
             code: product.code,
             price: Number.parseFloat(product.price),
             status: true,
-            stock: Number.parseFloat(product.stock),
+            stock: Number.parseInt(product.stock),
             category: product.category,
         }
     }
